@@ -26,59 +23,54 @@
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
-    "build": "ng build --prod",
+    "build": "ng build",
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e",
-    "buildProject": "npm install && npm prune && npm list --depth=0 && ng lint && ng test --single-run --code-coverage && ng e2e && ng build --deploy-url=odata-v4-ng && rm -R -f docs && mv dist docs",
+    "buildProject": "npm install && ng lint && ng test --watch=false --code-coverage && ng e2e && ng build --deploy-url=odata-v4-ng && rm -R -f docs && mv dist/odata-v4-ng docs",
     "packageProject": "ng-packagr -p ng-package.json",
     "publishDist": "npm publish dist",
     "buildPackagePublish": "npm run buildProject && npm run packageProject && npm run publishDist"
   },
   "private": false,
   "peerDependencies": {
-    "@angular/common": "^5.2.0",
-    "@angular/core": "^5.2.0",
-    "@angular/http": "^5.2.0",
-    "@angular/platform-browser": "^5.2.0",
-    "rxjs": "^5.5.6",
-    "zone.js": "^0.8.19",
-    "angular2-uuid": "^1.1.1"
+    "@angular/common": "^6.0.6",
+    "@angular/core": "^6.0.6",
+    "@angular/http": "^6.0.6",
+    "@angular/platform-browser": "^6.0.6",
+    "angular2-uuid": "^1.1.1",
+    "rxjs": "^6.0.0",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/animations": "^5.2.0",
-    "@angular/common": "^5.2.0",
-    "@angular/compiler": "^5.2.0",
-    "@angular/core": "^5.2.0",
-    "@angular/forms": "^5.2.0",
-    "@angular/http": "^5.2.0",
-    "@angular/platform-browser": "^5.2.0",
-    "@angular/platform-browser-dynamic": "^5.2.0",
-    "@angular/router": "^5.2.0",
-    "core-js": "^2.4.1",
-    "rxjs": "^5.5.6",
-    "zone.js": "^0.8.19",
-    "@angular/cli": "~1.7.3",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/language-service": "^5.2.0",
-    "@types/jasmine": "~2.8.3",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
-    "codelyzer": "^4.0.1",
-    "jasmine-core": "~2.8.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2",
-    "ts-node": "~4.1.0",
-    "tslint": "~5.9.1",
-    "typescript": "~2.5.3",
+    "@angular-devkit/build-angular": "~0.6.8",
+    "@angular/animations": "^6.0.6",
+    "@angular/cli": "~6.0.8",
+    "@angular/compiler": "^6.0.6",
+    "@angular/compiler-cli": "^6.0.6",
+    "@angular/forms": "^6.0.6",
+    "@angular/language-service": "^6.0.6",
+    "@angular/platform-browser-dynamic": "^6.0.6",
+    "@angular/router": "^6.0.6",
+    "@types/jasmine": "~2.8.6",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/node": "~8.9.4",
+    "codelyzer": "~4.2.1",
+    "core-js": "^2.5.4",
     "font-awesome": "^4.7.0",
-    "primeng": "^5.0.0",
-    "angular2-uuid": "^1.1.1",
-    "ng-packagr": "^2.2.0"
+    "jasmine-core": "~2.99.1",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~1.7.1",
+    "karma-chrome-launcher": "~2.2.0",
+    "karma-coverage-istanbul-reporter": "~2.0.0",
+    "karma-jasmine": "~1.1.1",
+    "karma-jasmine-html-reporter": "^0.2.2",
+    "ng-packagr": "^2.4.0",
+    "primeng": "^6.0.0",
+    "protractor": "~5.3.0",
+    "rxjs-tslint": "^0.1.5",
+    "ts-node": "~5.0.1",
+    "tslint": "~5.9.1",
+    "typescript": "~2.7.2"
   }
 }
